requirements.txt
fastapi
uvicorn
python-dotenv
langchain_groq
langchain_core

